requirements.txt
flask
flask-sqlalchemy
flask-admin
bootstrap-flask
flask-wtf
flask-login
mysqlclient
email-validator
wtforms-sqlalchemy